-Wsign-compare warnings fix, the warning is disabled, part 8(final)

+++ projects/miopen/driver/tensor_driver.hpp
@@ -78,7 +78,7 @@ inline void LengthReorder(std::vector<int>& lens, const std::initializer_list<in
     out_lens.reserve(indices.size());
     for(int index : indices)
     {
-        assert(0 <= index && index < lens.size());
+        assert(0 <= index && static_cast<size_t>(index) < lens.size());
         out_lens.push_back(std::move(lens[index]));
     }
     lens = std::move(out_lens);
